requirements.txt
alembic==1.13.1
asyncpg==0.29.0
psycopg-binary==3.1.10
SQLAlchemy==2.0.29
starlette==0.37.2
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.2
pydantic==2.7.1
pydantic-settings==2.2.1
uvicorn==0.29.0
psycopg==3.1.18
py-bcrypt==0.4
python-jose==3.3.0